Bug with name handling + temp code to change namespace
The name checks are still 'TODO', and some of the 'old' code was buggy. I consolidated the code and removed the bug. Propper checks are still 'TODO'.

I'm shifting pages from the 'Extension:' namespace into the main namespace, so I've temporarily added code to delete the page from the (phoney) 'Extension' namespace. This can be safely removed later.